+++ src/Commands.h
@@ -328,8 +328,9 @@ enum {
     CmdOpenSettingsFile = 522,
     CmdOpenFileWithSumatraFilePicker = 523,
     CmdSelectCurrentPage = 524,
+    CmdZoomFitVisible = 525,
 
-    CmdLast = 524,
+    CmdLast = 525,
     CmdFirstCustom = CmdLast + 100,
 
     // aliases, at the end to not mess ordering