+++ src/WindowTab.cpp
@@ -235,6 +235,9 @@ float WindowTab::NextToggleZoom() const {
         return kZoomFitContent;
     }
     if (kZoomFitContent == currZoom) {
+        return kZoomFitVisible;
+    }
+    if (kZoomFitVisible == currZoom) {
         return kZoomShrinkToFit;
     }
     return kZoomFitPage;